Crawler Manager refactored
