--- spider_manager/crawling_utils/crawling_utils/__init__.py
@@ -1,1 +1,0 @@
-from crawling_utils.crawling_utils import *
